Move ContextSampler and NegativeSampler to data/sampler.py
