+++ gluonnlp/data/sampler.py
@@ -19,11 +19,27 @@
 """Samplers. They define how the samples in a dataset will be iterated
 (e.g. in the order sorted by length). They can also be used to perform bucketing
 for speeding up the processing of variable-length sequences."""
-__all__ = ['SortedSampler', 'FixedBucketSampler', 'SortedBucketSampler']
+__all__ = [
+    'SortedSampler', 'FixedBucketSampler', 'SortedBucketSampler',
+    'ContextSampler', 'NegativeSampler'
+]
 
+import math
+import random
 import warnings
+
 import numpy as np
 from mxnet.gluon.data import Sampler
+
+try:
+    from numba import njit, prange
+    numba_njit = njit(nogil=True)
+except ImportError:
+    # Define numba shims
+    prange = range
+
+    def numba_njit(func):
+        return func
 
 
 def _match_bucket_keys(bucket_keys, seq_lengths):
@@ -323,3 +339,172 @@ class SortedBucketSampler(Sampler):
 
     def __len__(self):
         return (len(self._sort_keys) + self._batch_size - 1) // self._batch_size
+
+
+class NegativeSampler(Sampler):
+    """Sampler for drawing negatives from a smoothed unigram distribution.
+
+    Parameters
+    ----------
+    vocab : Vocab
+        The vocabulary specifies the set of tokens and their frequencies from
+        which the smoothed unigram distribution is computed.
+    batch_size : int
+        Maximum size of batches. Actual batch returned can be smaller when
+        running out of samples.
+    num_samples : int
+        Total number of samples that will be iterated over in batches.
+    negative : int, default 5
+        The number of negative samples to draw.
+    power : float, default 0.75
+        Smoothing factor.
+
+    """
+
+    def __init__(self, vocab, batch_size, num_samples, negative=5, power=0.75):
+        self.vocab = vocab
+        self.batch_size = batch_size
+        self.num_samples = num_samples
+        self.negative = negative
+        self.power = power
+
+        # Smoothed unigram counts for negative sampling. Negatives can be drawn
+        # by sampling a number in [0, self._smoothed_cumsum[-1]) and finding
+        # the respective index with np.searchsorted.
+        self._smoothed_token_freq_cumsum = np.cumsum((np.array(
+            vocab.idx_to_counts)**self.power).astype(np.int))
+
+    def __len__(self):
+        return math.ceil(self.num_samples / float(self.batch_size))
+
+    def __iter__(self):
+        def _generator():
+            pointer = 0
+            while True:
+                size = min(self.batch_size, self.num_samples - pointer)
+                yield np.searchsorted(self._smoothed_token_freq_cumsum,
+                                      np.random.randint(
+                                          self._smoothed_token_freq_cumsum[-1],
+                                          size=(size,
+                                                self.negative)))
+                pointer += size
+
+                if pointer == self.num_samples:
+                    break
+
+        return _generator()
+
+
+class ContextSampler(Sampler):
+    """Sample batches of contexts (and their masks) from a corpus.
+
+    The context size is choosen uniformly at random for every sample from [1,
+    `window`]. The mask is used to mask entries that lie outside of the
+    randomly chosen context size. Contexts do not cross sentence boundaries.
+
+    Batches are created lazily, to avoid generating all batches for shuffling
+    before training, simply shuffle the dataset before passing it to the
+    ContextSampler.
+
+    Parameters
+    ----------
+    coded : list of lists of int
+        List of coded sentences. A coded sentence itself is a list of token
+        indices. Context samples do not cross sentence boundaries.
+    batch_size : int
+        Maximum size of batches. Actual batch returned can be smaller when
+        running out of samples.
+    window : int, default 5
+        The maximum context size.
+
+    Attributes
+    ----------
+    num_samples : int
+        Overall number of samples that are iterated over in batches. This is
+        the total number of token indices in `coded`.
+
+    """
+
+    def __init__(self, coded, batch_size, window=5):
+        self.batch_size = batch_size
+        self.window = window
+        coded = [c for c in coded if len(c) > 1]
+        self._sentence_boundaries = np.cumsum([len(s) for s in coded])
+        self.num_samples = self._sentence_boundaries[-1]
+        self._coded = np.concatenate(coded)
+
+    def __len__(self):
+        return math.ceil(self.num_samples / float(self.batch_size))
+
+    def __iter__(self):
+        return _context_generator(self._coded, self._sentence_boundaries,
+                                  self.window, self.batch_size)
+
+
+@numba_njit
+def _get_sentence_start_end(sentence_boundaries, sentence_pointer):
+    end = sentence_boundaries[sentence_pointer]
+    if sentence_pointer == 0:
+        start = 0
+    else:
+        start = sentence_boundaries[sentence_pointer - 1]
+    return start, end
+
+
+@numba_njit
+def _context_generator(coded_sentences, sentence_boundaries, window,
+                       batch_size):
+    word_pointer = 0
+    while True:
+        batch_size = min(batch_size, len(coded_sentences) - word_pointer)
+        center = coded_sentences[word_pointer:
+                                 word_pointer + batch_size].astype(np.float32)
+        context = np.zeros((batch_size, window * 2), dtype=np.float32)
+        mask = np.zeros((batch_size, window * 2), dtype=np.float32)
+
+        for i in prange(batch_size):
+            context_, mask_ = _get_context(word_pointer + i, coded_sentences,
+                                           sentence_boundaries, window)
+            context[i] = context_
+            mask[i] = mask_
+
+        word_pointer += batch_size
+
+        yield center, context, mask
+
+        if word_pointer >= sentence_boundaries[-1]:
+            break
+
+
+@numba_njit
+def _get_context(word_pointer, coded_sentences, sentence_boundaries, window):
+    sentence_pointer = np.searchsorted(sentence_boundaries, word_pointer)
+    sentence_start, sentence_end = _get_sentence_start_end(
+        sentence_boundaries, sentence_pointer)
+
+    random_window_size = random.randint(1, window)
+
+    start_idx = max(sentence_start, word_pointer - random_window_size)
+    # First index outside of the window
+    end_idx = min(sentence_end, word_pointer + random_window_size + 1)
+
+    # A random reduced window size is drawn. The mask masks entries
+    # that fall inside the window, but outside random the reduced
+    # window size.
+    mask = np.ones(window * 2, dtype=np.float32)
+    mask[end_idx - start_idx:] = 0
+    context = np.zeros(window * 2, dtype=np.float32)
+
+    # Get contexts
+    next_context_idx = 0
+    context[:word_pointer - start_idx] = \
+        coded_sentences[start_idx:word_pointer]
+    next_context_idx += word_pointer - start_idx
+    context[next_context_idx:next_context_idx + end_idx -
+            (word_pointer + 1)] = coded_sentences[word_pointer + 1:end_idx]
+    next_context_idx += end_idx - (word_pointer + 1)
+
+    # Set mask
+    mask[next_context_idx:] = 0
+
+    return context, mask
